+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Owner check NFT on bridge
 - No data message on staking tab
 - Open graph meta information
+- Data to connect the network in metamask
 
 #### Changed
 - Version api packages